backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
sqlalchemy
psycopg[binary]
langgraph-checkpoint-redis
redis

requests
yfinance

langchain-core
deepagents
langchain-openai
